+++ src/conditions/LoginAttemptsCondition.php
@@ -6,6 +6,7 @@ use Craft;
 use craft\elements\conditions\ElementCondition;
 
 use amici\LoginAttempts\conditions\rules\LoginStatusConditionRule;
+use amici\LoginAttempts\conditions\rules\LoginTypeConditionRule;
 use amici\LoginAttempts\conditions\rules\IpAddressConditionRule;
 use amici\LoginAttempts\conditions\rules\UserConditionRule;
 
@@ -18,6 +19,7 @@ class LoginAttemptsCondition extends ElementCondition
     {
         $types = array_merge(parent::selectableConditionRules(), [
             LoginStatusConditionRule::class,
+            LoginTypeConditionRule::class,
             IpAddressConditionRule::class,
             UserConditionRule::class,
         ]);